Pacman Main Code
The classical single player pacman game. The player will move pacman using WAXD keys on the keyboard. The board will be drawn as a fixed maze with pILLS in it and the prison at the middle which keeps the ghosts. The game will have 5 ghosts and the pacman. At the game start, the ghosts will initially be trapped in the prison for 10 seconds after which they will exit the prison. The pacman will start from the middle of the game board. There will be 4 power-ups at the four corners of the game board. If the pacman eats those power-ups, it will be invincible for 10 seconds. Any ghost it eats while it is invincible will be teleported back to the prison in the middle of the maze.